fix: properly give Gio OS main thread
This commit demonstrates a change that may fix the Gio-
based previewing on macOS. It restructures the main
function so that Gio's application main thread gains
control of the OS main thread, and the previous logic
is instead executed in a goroutine.

I don't have an easy way to test this on a mac, so I
cannot be certain this will fix your problems. I also
don't know if my changes broke any important CLI behaviors.
Please double-check me on all of the changes I made.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>