gui: construct layout.Context only once per frame
Gio interfaces are described in terms of an operation list of type
op.Ops. This operation list contains both the graphical and input
description of your GUI interface. To make this operation list
easier to use, we typically wrap it in a layout.Context, which
provides information like the available screen space and pixel
density.

This commit fixes a problem in which both th emain window event
loop and the draw() function constructed a layout.Context from
the same operation list. When you construct a context, it clears
the contents of the operation list in preparation for drawing a
new frame. This meant that the key input operations inserted into
the operation list in the main event loop were being cleared by
the draw() function and thus never processed by Gio.

While here, I simplified some Gio usage. It isn't necessary to pass
the window and frame event to UI functions, as you can trigger
invalidations using a Gio op.InvalidateOp (which has less latency
than window.Invalidate()) and you can get the available screen space
from the constraints on the layout.Context.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>